+++ Assets/RVO2/AgentData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6d252b606a986d346a3091cf9fbd2446
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
