+++ Packages/manifest.json
@@ -1,9 +1,9 @@
 {
   "dependencies": {
-    "com.unity.burst": "1.6.6",
-    "com.unity.collections": "1.4.0",
-    "com.unity.ide.rider": "3.0.16",
-    "com.unity.ide.visualstudio": "2.0.16",
+    "com.unity.burst": "1.8.12",
+    "com.unity.collections": "2.4.2",
+    "com.unity.ide.rider": "3.0.27",
+    "com.unity.ide.visualstudio": "2.0.22",
     "com.unity.ide.vscode": "1.2.5",
     "com.unity.mathematics": "1.2.6",
     "com.unity.test-framework": "1.1.33",
