+++ ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2020.3.41f1
-m_EditorVersionWithRevision: 2020.3.41f1 (7c19dc9acfda)
+m_EditorVersion: 2022.3.18f1
+m_EditorVersionWithRevision: 2022.3.18f1 (d29bea25151d)
